Solr install scripted (along with zookeeper) - not test yet though - but should work ;-)
So I brought the solr portion of the script up, on par with the zookeeper portion.
The test_solr function needs to be looked at more closely and implemented
The whole tomcat business is still commented out and needs to be looked at

There are a few NOTEs in the script indicating the preferred use of
the checked_get function, but the inability to use it just yet because
it needs to be factored out of the uber script along with other
utility functions.

There are also testing things to do with regards to timing between
zookeeper and solr booting.  Also the boot order is still in question.
right now it is zookeeper then solr on start and reverse on shutdown.

But right now I am bashed out and wanna go partying! TGIF!
-Gavin :-)