Solr installation and configuration more complete
Able to clean up the installation process.  Now everything gets setup
(installed) and configured "properly".

Because there is not a notion of version that is inherent in the
distribution we had to tack in on by using the version number in the
naming of the distribuiton tar.gz file.  That version value is now
writen to a VERSION file at the top level solr installation directory.
The version check function looks at the VERSION file and compares that
the the version that the script is set to install.

Took out some superfluous output.

TODO... the test function for solr