Made zookeeper configuration less destructive
Will only go to the net (distribution server) to pull down zookeeper
configuration file (zoo.cfg) when the "dataDir" key/value property is
not present or if the --force flag is used.  Otherwise it will read
the local zoo.cfg file and make the key/value replacement for
dataDir=${zookeeper_data_dir}.  When the config file is pulled from
the distribution server all key/value properties in the base
distribution configuration file are used and will usurp those in the
local file.  Additional properties will not be modified.

This means that the distribution server contains the proper minimum
set of properties for zookeeper configuration.  To get a clean file
from the distribution server the user will have to remove the zoo.cfg
file, so that it can be replaced with the base file from the
distribution server.

Because of the editing of the file it is no longer necessary to grep
for the value of dataDir to replace since we are basing the
replacement on the KEY not the value.