Bug Fix for mac... md5sum work around...
The md5sum program is not natively on macs.
There is an md5 program though.

So this fix involved writing our own internal md5sum function that
tests for "md5sum" and if not uses "md5" (which is natively on macs)
to calculate the checksum.

This fixes the infinite loop bug that showed up.

caveat, this fix may not be portable to all platforms.  At the moment
is tested for mac os x and rh/centos unix platforms.