A bit of os detection to help select how to get mod time for files
Wrapping the call to stat in a "private" function get_mod_time_ and
taking advantage of the check_os logic, we can select the correct call
to get the modification times of files.

At the moment we only look at vanilla Linux vs Darwin operating
systems (based on the uname command output).

P.S.
Decided not to use the check_os logic in the md5sum detection (the
last commit) because the md5sum issue is not structly a Linux vs
Darwin issue (AFAIK) - so let's keep that test more open and not
restricted by operating system discriminiation.