Deduping cache with sort and awk - tac is sexy but too esoteric
Using sort and awk (and sort again) we dedup the cache of previously
downloaded files.  Note: the entire list is loaded into memory for
sort and awk to do their jobs.  Given how much memory computers have
these days and given that a download script should not be downloading
millions of files and given that the operation, though in-memory, is
ephemeral, I think that this solution should be fine.

-gavin