Looking for main distribution location or devel depending...
It makes sense that the third party libraries be in a static location,
because they are not under our (esgf) development. Hmmm... I waver
with this back and forth... But for now, I think it is most imporant
to keep process exceptions to a minimum.  So for the moment let's go
back to treating all things in either master or devel regimes.

-gavin