Move old style facets.properties file out of the way
In effort to address the support issue of having old facet files in
place.  Now we look for the first line - version tag - as a clue that the
file is indeed up to date.  If it is an old style file then it is
moved aside and the default one is put in place.