Double the wait time - in response to Prashanth's observation on his node
So now I wait a minute instead of 30 seconds.... This is not an error,
but I don't like posting scary [FAIL] messages that are inaccurate.

3.4	Issue with Apache Solr startup

During the startup, the Apache solr service is reported to have failed
to start. This is generally due to the script not waiting long enough
for the service to start. To confirm, wait another minute or so and do
netstat -an|grep 894 The solr services listen on 8983 and 8984
ports. If they are listening, everything is just fine.

(maybe even a minute is not long enough? - Let's see)